Clean up code and captital basic type for enum

+++ schema/parser/enum_test.go
@@ -5,30 +5,30 @@ import "testing"
 func TestEnumParser(t *testing.T) {
 	runTests(t, TestCases{
 		{
-			Input:  `enum A int64 {}`,
-			Output: `enum A int64 {}`,
+			Input:  `enum A Int64 {}`,
+			Output: `enum A Int64 {}`,
 		},
 		{
-			Input: `enum A int64 { A = 1 }`,
+			Input: `enum A Int64 { A = 1 }`,
 			Output: `
-enum A int64 {
+enum A Int64 {
 	A = 1
 }`,
 		},
 		{
-			Input: `enum A int64 { A = 1 B C }`,
+			Input: `enum A Int64 { A = 1 B C }`,
 			Output: `
-enum A int64 {
+enum A Int64 {
 	A = 1
 	B
 	C
 }`,
 		},
 		{
-			Input: `enum A int64 { A = 1 B 
+			Input: `enum A Uint8 { A = 1 B 
 				C = 4 }`,
 			Output: `
-enum A int64 {
+enum A Uint8 {
 	A = 1
 	B
 	C = 4
